Fixed ableton special 8-bit value, values going from -1 to 255

+++ service/internal/indexer/mapping_audio_track.go
@@ -19,7 +19,7 @@ type AudioTrackDocument struct {
 	MemorizedFirstClipName string `json:"memorizedFirstClipName,omitempty"`
 	Annotation             string `json:"annotation,omitempty"`
 
-	Color uint8 `json:"color,omitempty"`
+	Color int16 `json:"color,omitempty"`
 }
 
 func NewAudioTrackDocument() *AudioTrackDocument {
