+++ service/internal/indexer/mapping_midi_track.go
@@ -19,7 +19,7 @@ type MidiTrackDocument struct {
 	MemorizedFirstClipName string `json:"memorizedFirstClipName,omitempty"`
 	Annotation             string `json:"annotation,omitempty"`
 
-	Color uint8 `json:"color,omitempty"`
+	Color int16 `json:"color,omitempty"`
 }
 
 func NewMidiTrackDocument() *MidiTrackDocument {
